+++ src/main/java/lotto/util/Rank.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.util;
 
 public enum Rank {
     FIRST,
