+++ src/test/java/lotto/domain/MatchNumberTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.util.MatchResult;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
