+++ src/test/java/lotto/domain/MoneyTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.*;
@@ -12,6 +13,7 @@ public class MoneyTest {
     }
 
     @Test
+    @DisplayName("잘못된 금액을 입력하면 예외 발생.")
     public void exceptionTest(){
         assertThatExceptionOfType(IllegalArgumentException.class).isThrownBy(()->{
             new Money(999);
@@ -20,6 +22,7 @@ public class MoneyTest {
     }
 
     @Test
+    @DisplayName("구매가능한 로또 개수를 반환한다.")
     public void howManyTest() {
         Money money = new Money(14000);
         assertThat(money.howMany()).isEqualTo(14);
