--- src/test/java/lotto/domain/RankingsDtoTest.java
@@ -1,6 +1,0 @@
-package lotto.domain;
-
-import org.junit.jupiter.api.Test;
-
-public class RankingsDtoTest {
-}
