+++ src/test/java/lotto/util/RandomNumberTest.java
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class RandomNumberTest {
+    final int NUM_OF_REPEAT = 100;
+
     RandomNumberGenerator randomNumberGenerator;
     @BeforeEach
     void setUp(){
@@ -14,11 +16,15 @@ public class RandomNumberTest {
 
     @Test
     void rangeTest(){
-        randomNumberGenerator.getNumbers().stream().forEach(num -> assertThat(num).isBetween(1,45));
+        for (int i = 0; i < NUM_OF_REPEAT; i++) {
+            randomNumberGenerator.getNumbers().stream().forEach(num -> assertThat(num).isBetween(1, 45));
+        }
     }
 
     @Test
     void duplicationTest(){
-        assertThat(randomNumberGenerator.getNumbers()).doesNotHaveDuplicates();
+        for (int i = 0; i < NUM_OF_REPEAT; i++) {
+            assertThat(randomNumberGenerator.getNumbers()).doesNotHaveDuplicates();
+        }
     }
 }
